Jour2/ApplicationReconversion/requirements.txt
streamlit>=1.29.0
openai>=1.6.0
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.18.0
streamlit-option-menu>=0.3.6
